+++ high-concurrency-processor/src/main/java/jeff/highconcurrency/controller/HighConcurrencyController.java
@@ -1,13 +1,19 @@
 package jeff.highconcurrency.controller;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.sun.org.apache.xpath.internal.operations.Bool;
+import jeff.common.consts.DemoMember;
+import jeff.common.entity.bo.MyRequestContext;
 import jeff.common.entity.dto.send.ResponseObject;
+import jeff.common.exception.MyException;
+import jeff.common.util.LogUtil;
 import jeff.highconcurrency.persistent.model.mongo.dao.ReactiveFlashSaleEventLogRepo;
 import jeff.highconcurrency.service.FlashSaleEventService;
 import jeff.highconcurrency.util.redis.util.MyReactiveRedisUtil;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Mono;
 
 /**
@@ -18,6 +24,7 @@ import reactor.core.publisher.Mono;
  *
  * 當使用Mono和Flux來處理請求時，WebFlux不會為每個請求分配獨立的執行緒(不同於以往的同步模型)，而是通過非同步處理，更高效地利用執行緒和資源。
  */
+@Slf4j
 @RestController
 public class HighConcurrencyController {
 
@@ -29,6 +36,9 @@ public class HighConcurrencyController {
 
     @Autowired
     MyReactiveRedisUtil myReactiveRedisUtil;
+
+    @Autowired
+    LogUtil logUtil;
 
     @GetMapping("/hello")
     public Mono hello() {
@@ -68,4 +78,20 @@ public class HighConcurrencyController {
     public Mono<Void> test5() {
         return myReactiveRedisUtil.removeAllKeys();
     }
+
+    /**
+     * 客戶端快閃銷售案件的商品下單時的接口。
+     * 通常是搶門票的頁面，進入快閃銷售案件特有的結帳頁面後，按下送出所打的API。
+     */
+    @PostMapping("/order/flash")
+    public Mono<ResponseEntity<ResponseObject>> createAnOrderFromFlashSalesEvent(@RequestBody JsonNode param) throws MyException {
+        System.out.println("222" + param.toString());
+//        myRequestContext.setAuthenticatedMemberId(DemoMember.CUSTOMER.getId()); // TODO 此API的請求者就是買家，目前先寫死，所以前端也不用傳這個參數
+
+        System.out.println("222" + param.toString());
+
+//        Mono.just(ResponseEntity.ok(flashSaleEventService.consumeFlashSaleEvent()))
+        return Mono.empty();
+    }
+
 }