+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - core-processor新增完成訂單的接口
+- high-concurrency-processor完成過濾鏈撰寫
 
 ## [0.3.1] - 2024-01-14